Clean up descartes planner
