+++ tesseract/tesseract_planning/tesseract_motion_planners/cmake/tesseract_motion_planners-config.cmake.in
@@ -13,6 +13,7 @@ find_dependency(descartes_opw)
 find_dependency(descartes_tesseract)
 find_dependency(tesseract)
 find_dependency(tesseract_common)
+find_dependency(console_bridge)
 find_dependency(Boost)
 
 include("${CMAKE_CURRENT_LIST_DIR}/@PROJECT_NAME@-targets.cmake")
